requirements.txt
requests
dnspython
python-whois
cryptography
pandas
beautifulsoup4
colorama
python-dotenv
